fixed somethings
