+++ src/components/core/BaseLayout/BaseLayout.jsx
@@ -3,40 +3,13 @@ import { useParams, useNavigate, Navigate } from "react-router-dom";
 
 import "./BaseLayout.css";
 
-import Main from "../../pages/Main/Main";
-import Dashboard from "../../pages/Dashboard/Dashboard";
-import EditBoard from "../../pages/EditBoard/EditBoard";
 import Board from "../../pages/Board/Board";
 import NotFound from "../../pages/NotFound/NotFound";
-import Sidebar from "../Sidebar/Sidebar";
 import DatabaseContext from "../../../contexts/DatabaseContext";
+import MainLayout from "../MainLayout/MainLayout";
+import SideLayout from "../SideLayout/SideLayout";
 
 let boards;
-
-function mainLayout() {
-	return (
-		<div className="main_layout">
-			<Main />
-		</div>
-	);
-}
-
-function SideLayout(currBoard, page) {
-	return (
-		<div className="side_layout">
-			<div className="side_layout__sidebar">
-				<Sidebar />
-			</div>
-			<div className="side_layout__main">
-				{page === "dashboard" ? (
-					<Dashboard board={currBoard} />
-				) : (
-					<EditBoard board={currBoard} />
-				)}
-			</div>
-		</div>
-	);
-}
 
 function BoardLayout() {
 	return (
@@ -74,8 +47,7 @@ function BaseLayout({ page, ...props }) {
 	}
 	//Render the current board into the page
 	if (page === "main") {
-		// return <Main {...props} />;
-		return mainLayout();
+		return MainLayout();
 	} else if (page === "dashboard") {
 		return SideLayout(currBoard, "dashboard");
 	} else if (page === "edit") {
