+++ src/components/core/MainLayout/MainLayout.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+import Main from "../../pages/Main/Main";
+
+function MainLayout() {
+	return (
+		<div className="main_layout">
+			<Main />
+		</div>
+	);
+}
+
+export default MainLayout;
