+++ src/components/core/Sidebar/Sidebar.css
@@ -4,7 +4,7 @@
 	display: flex;
 	flex-direction: column;
 	font-size: large;
-	position: fixed;
+	position: sticky;
 }
 
 .sidebar__title {
